refactor(AddSectionPolicy): rename
- 수행하는 역할에 어울리도록 좀 더 좁은 범위의 이름으로 변경
- 기존의 이름(AddSectionPolicy)는 상행종점역과 하행 종점역을 고려해서 더 범용적인 오브젝트에서 사용할 예정
